:sparkles: api: splitting certification results csv generation
* Certification csv are now independent from each other in their construction
