+++ api/lib/infrastructure/utils/csv/certification-results/SessionCertificationResultsCsvBuilder.js
