+++ api/scripts/certification/generate-certification-csv-results-by-session-ids.js
@@ -10,7 +10,6 @@ import bluebird from 'bluebird';
 import lodash from 'lodash';
 const { isEmpty } = lodash;
 import { logger } from '../../lib/infrastructure/logger.js';
-import * as certificationResultUtils from '../../lib/infrastructure/utils/csv/certification-results.js';
 import { usecases } from '../../lib/domain/usecases/index.js';
 import { temporaryStorage } from '../../lib/infrastructure/temporary-storage/index.js';
 import { disconnect } from '../../db/knex-database-connection.js';
@@ -39,7 +38,7 @@ async function main() {
       return;
     }
 
-    const csvResult = await certificationResultUtils.getSessionCertificationResultsCsv({
+    const csvResult = await usecases.getSessionCertificationResultsCsv({
       session,
       certificationResults,
     });
