+++ api/tests/unit/infrastructure/utils/csv/certification-results/CertificationResultsCsvValues_test.js
@@ -1,7 +1,7 @@
-import { expect, domainBuilder } from '../../../../test-helper.js';
-import { getI18n } from '../../../../tooling/i18n/i18n.js';
-import { CertificationResult } from '../../../../../lib/domain/models/CertificationResult.js';
-import { CertificationResultsCsvValues } from '../../../../../lib/infrastructure/utils/csv/CertificationResultsCsvValues.js';
+import { expect, domainBuilder } from '../../../../../test-helper.js';
+import { getI18n } from '../../../../../tooling/i18n/i18n.js';
+import { CertificationResult } from '../../../../../../lib/domain/models/CertificationResult.js';
+import { CertificationResultsCsvValues } from '../../../../../../lib/infrastructure/utils/csv/certification-results/CertificationResultsCsvValues.js';
 
 describe('Unit | Infrastructure | Utils | Csv | CertificationResultsCsvValues', function () {
   let i18n;
